feat: 사전 식물 상세 정보 SSG 적용
