+++ src/components/@common/BackButton/BackButton.style.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Center = styled.button`
+  display: flex;
+  align-items: center;
+`;
