+++ src/components/Toast/ToastList.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { memo } from 'react';
 import { createPortal } from 'react-dom';
 import { useRecoilValue } from 'recoil';
